requirements.txt
langchain
langchain-groq
python-dotenv
speech-recognition
pyttsx3